backend/requirements.txt
fastapi
uvicorn
python-multipart
qdrant-client
elasticsearch
torch
torchvision
transformers
pillow
